Made a minimal test for selecting multiple items and applying tags.
Very minimal test created.
This makes a public static variable from MainActivity in order to check when the items are fetched from firestore. (which is obviously a bad practice) Animations must be disabled in Android setting for this test to work. Might be possible to implement it to automatically disable animations using uiautomator (haven't done this yet). Can make tests more extensive later...